# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/expenses": {
      "post": {
        "operationId": "ExpenseController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpenseRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new expense",
        "tags": [
          "Expense"
        ]
      },
      "get": {
        "operationId": "ExpenseController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "storeId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "familyId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PERSONAL",
                "FAMILY"
              ]
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "valueMin",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "valueMax",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedExpenseResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all expenses with pagination and filters",
        "tags": [
          "Expense"
        ]
      }
    },
    "/expenses/statistics": {
      "get": {
        "operationId": "ExpenseController_getStatistics",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "storeId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "familyId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PERSONAL",
                "FAMILY"
              ]
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "valueMin",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "valueMax",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseStatisticsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get expense statistics",
        "tags": [
          "Expense"
        ]
      }
    },
    "/expenses/trends": {
      "get": {
        "operationId": "ExpenseController_getTrends",
        "parameters": [
          {
            "name": "dateFrom",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "day",
              "type": "string",
              "enum": [
                "day",
                "week",
                "month"
              ]
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "storeId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "familyId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PERSONAL",
                "FAMILY"
              ]
            }
          },
          {
            "name": "valueMin",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "valueMax",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxLabels",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 2,
              "maximum": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExpenseTrendPointResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get expense trends over time",
        "tags": [
          "Expense"
        ]
      }
    },
    "/expenses/{id}": {
      "get": {
        "operationId": "ExpenseController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get an expense by ID",
        "tags": [
          "Expense"
        ]
      },
      "put": {
        "operationId": "ExpenseController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExpenseRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an expense",
        "tags": [
          "Expense"
        ]
      },
      "delete": {
        "operationId": "ExpenseController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an expense",
        "tags": [
          "Expense"
        ]
      }
    },
...
  },
  "info": {
    "title": "Fynans API",
    "description": "Fynans personal finance API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TransactionUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "image": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName"
        ]
      },
      "TransactionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "type": {
            "enum": [
              "EXPENSE",
              "INCOME"
            ],
            "type": "string"
          },
          "scope": {
            "enum": [
              "PERSONAL",
              "FAMILY"
            ],
            "type": "string"
          },
          "value": {
            "type": "number"
          },
          "recordedAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/TransactionUserDto"
          }
        },
        "required": [
          "id",
          "userId",
          "type",
          "scope",
          "value",
          "recordedAt",
          "createdAt",
          "updatedAt",
          "user"
        ]
      },
...
      "ExpenseCategoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "parentId": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "isConnectedToStore": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "parentId",
          "name",
          "createdAt",
          "updatedAt",
          "isConnectedToStore"
        ]
      },
...
      "CreateExpenseItemRequestDto": {
        "type": "object",
        "properties": {
          "sizeUnit": {
            "type": "string",
            "enum": [
              "kg",
              "g",
              "l",
              "ml",
              "cl"
            ]
          },
          "expenseId": {
            "type": "string",
            "format": "uuid"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "itemName": {
            "type": "string"
          },
          "itemPrice": {
            "type": "number",
            "minimum": 0
          },
          "discount": {
            "type": "number",
            "minimum": 0
          },
          "quantity": {
            "type": "number",
            "minimum": 0.001
          },
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "sizeValue": {
            "type": "number",
            "minimum": 0.001
          }
        },
        "required": [
          "expenseId",
          "categoryId",
          "itemName",
          "itemPrice"
        ]
      },
      "ExpenseItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "itemId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "expenseId": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "discount": {
            "type": "number"
          },
          "finalPrice": {
            "type": "number"
          },
          "discountPercentage": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "itemId",
          "name",
          "expenseId",
          "categoryId",
          "price",
          "discount",
          "finalPrice",
          "discountPercentage",
          "quantity",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateExpenseRequestDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateExpenseItemRequestDto"
            }
          },
          "scope": {
            "type": "string",
            "enum": [
              "PERSONAL",
              "FAMILY"
            ]
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "storeName": {
            "type": "string"
          },
          "storeLocation": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "minimum": 0
          },
          "note": {
            "type": "string"
          },
          "familyId": {
            "type": "string",
            "format": "uuid"
          },
          "recordedAt": {
            "type": "string"
          }
        },
        "required": [
          "categoryId"
        ]
      },
      "StoreResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "location",
          "createdAt",
          "updatedAt"
        ]
      },
      "ExpenseResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "transactionId": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "storeId": {
            "type": "string",
            "nullable": true
          },
          "transaction": {
            "$ref": "#/components/schemas/TransactionResponseDto"
          },
          "category": {
            "$ref": "#/components/schemas/ExpenseCategoryResponseDto"
          },
          "store": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/StoreResponseDto"
              }
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpenseItemResponseDto"
            }
          },
          "matchedItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpenseItemResponseDto"
            }
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "transactionId",
          "categoryId",
          "storeId",
          "transaction",
          "category",
          "store",
          "items",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedExpenseResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpenseResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit"
        ]
      },
      "ExpenseByCategoryDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string"
          },
          "categoryName": {
            "type": "string"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "categoryId",
          "categoryName",
          "total"
        ]
      },
      "ExpenseByStoreDto": {
        "type": "object",
        "properties": {
          "storeId": {
            "type": "string"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "storeId",
          "total"
        ]
      },
      "ExpenseStatisticsResponseDto": {
        "type": "object",
        "properties": {
          "totalExpenses": {
            "type": "number"
          },
          "expenseCount": {
            "type": "number"
          },
          "averageExpense": {
            "type": "number"
          },
          "expensesByCategory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpenseByCategoryDto"
            }
          },
          "expensesByStore": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpenseByStoreDto"
            }
          }
        },
        "required": [
          "totalExpenses",
          "expenseCount",
          "averageExpense",
          "expensesByCategory",
          "expensesByStore"
        ]
      },
      "ExpenseTrendPointResponseDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "total": {
            "type": "number"
          },
          "count": {
            "type": "number"
          },
          "showLabel": {
            "type": "boolean"
          }
        },
        "required": [
          "date",
          "total",
          "count",
          "showLabel"
        ]
      },
      "UpdateExpenseRequestDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "storeId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    }
  }
}